Feat: view object and better sticky spans
Adding a view that sits between data and builder. Point is to make it possible to filter and sort the data. Filters have been added to the builder but do not yet work.

Sticky group labels have also been improved. Column group labels stick to the left and index group labels stick to the top.